+++ grammars/hujson/bindings/python/tree_sitter_hujson/__init__.py
@@ -1,5 +1,0 @@
-"Hujson grammar for tree-sitter"
-
-from ._binding import language
-
-__all__ = ["language"]